+++ frontend/src/pages/CategoryPage/CategoryPage.jsx
@@ -13,8 +13,6 @@ const CategoryPage = () => {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
 
-  const [relatedPins, setRelatedPins] = useState([]);
-
   useEffect(() => {
     const loadCategoryAndPins = async () => {
       try {
