+++ frontend/src/services/pinService.js
@@ -1,8 +1,9 @@
 import axios from "axios";
 
 const API_BASE_URL = "http://localhost:4000/api";
+const DEFAULT_PAGE_LIMIT = 10;
 
-export const fetchPins = async (page = 1, limit = 10) => {
+export const fetchPins = async (page = 1, limit = DEFAULT_PAGE_LIMIT) => {
   try {
     const response = await axios.get(`${API_BASE_URL}/pins`, {
       params: { page, limit }
